change how code is organized, change router and queryClient location, begin rendering data from boxscore
