+++ package.json
@@ -25,7 +25,8 @@
     "react-day-picker": "^9.5.1",
     "react-dom": "^18.3.1",
     "react-nba-logos": "^1.1.3",
-    "react-router": "^7.1.5"
+    "react-router": "^7.1.5",
+    "react-router-dom": "^7.1.5"
   },
   "devDependencies": {
     "@eslint/js": "^9.17.0",
