+++ src/components/DatePicker.tsx
@@ -1,12 +1,16 @@
 import {useState} from "react";
 import {format} from "date-fns";
 import {DayPicker} from "react-day-picker";
+import type {SetURLSearchParams} from "react-router";
+import {useSearchParams} from "react-router";
 
 
 
-function DatePicker({setSearchParams}: SetURLSearchParams) {
+function DatePicker() {
   const [calendarOpen, setCalendarOpen] = useState(false);
   const [dateSelected, setDateSelected] = useState<Date>();
+  const [searchParams, setSearchParams] = useSearchParams({date: ""});
+
 
   return (
     <DayPicker
