+++ src/main.tsx
@@ -1,13 +1,22 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App'
-import {BrowserRouter} from "react-router";
+import {StrictMode} from "react";
+import {createRoot} from "react-dom/client";
+import "./index.css";
+import App from "./App";
+import Boxscore from "./routes/games/boxscore/Boxscore";
+import {BrowserRouter, Routes, Route} from "react-router";
+import {QueryClient, QueryClientProvider} from "@tanstack/react-query";
+
+const queryClient = new QueryClient();
 
 createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <BrowserRouter>
-      <App />
-    </BrowserRouter>
+    <QueryClientProvider client={queryClient}>
+      <BrowserRouter>
+        <Routes>
+          <Route path="/" element={<App />} />
+          <Route path="games/:gameId/boxscore" element={<Boxscore />} />
+        </Routes>
+      </BrowserRouter>
+    </QueryClientProvider>
   </StrictMode>
 );
